+++ faux_macros/src/self_type.rs
@@ -28,22 +28,6 @@ impl SelfType {
             SelfType::Arc => Some(quote!(std::sync::Arc)),
         }
     }
-
-    pub fn from_path(type_path: &syn::TypePath) -> Self {
-        let segment = type_path.path.segments.last().unwrap();
-        let ident = &segment.ident;
-
-        // can't match on Ident
-        if ident == "Rc" {
-            SelfType::Rc
-        } else if ident == "Arc" {
-            SelfType::Arc
-        } else if ident == "Box" {
-            SelfType::Box
-        } else {
-            SelfType::Owned
-        }
-    }
 }
 
 impl Default for SelfType {